feat(mail): implement mail module with multi-account support and inbox functionality
- Add EmptyMailGuide component for guiding users to configure their email accounts.
- Create MailAccountTabs component for switching between multiple email accounts.
- Implement MailHeaderActions component for settings and refresh actions in the mail header.
- Develop MailInboxList component to display the list of emails in the inbox.
- Add MailReader component to render the selected email content.
- Create MailTab component as the main entry point for the mail functionality, integrating all components.
- Introduce mail configuration constants and utility functions for handling email operations.
- Implement useMail hook for managing mail state, fetching inbox data, and handling account switching.
- Define TypeScript types for mail-related data structures and component props.
