+++ src/renderer/components/states/maxExpand/MaxExpandContentLazy.tsx
@@ -35,7 +35,7 @@ const UrlFavoritesTab = lazy(() => import('./components/UrlFavoritesTab').then((
 const LocalFileSearchTab = lazy(() => import('./components/localFileSearch/components/LocalFileSearchTab').then((module) => ({ default: module.LocalFileSearchTab })));
 const ClipboardHistoryTab = lazy(() => import('./components/clipBoardHistory').then((module) => ({ default: module.ClipboardHistoryTab })));
 const AlbumTab = lazy(() => import('./components/album/components/AlbumTab').then((module) => ({ default: module.AlbumTab })));
-const MailTab = lazy(() => import('./components/MailTab').then((module) => ({ default: module.MailTab })));
+const MailTab = lazy(() => import('./components/mail').then((module) => ({ default: module.MailTab })));
 const SettingsTab = lazy(() => import('./components/SettingsTab').then((module) => ({ default: module.SettingsTab })));
 const CountdownTab = lazy(() => import('./components/countdown').then((module) => ({ default: module.CountdownTab })));
 const MemoTab = lazy(() => import('./components/memo/components/MemoTab').then((module) => ({ default: module.MemoTab })));
